refactor: move columns to simpler namespace
